add coffee_id reference to rating schema

+++ coffee_backend/schemas/rating.py
@@ -12,7 +12,7 @@ class Rating(BaseModel):
         description="The id of the rating",
         example=UUID("123e4567-e89b-12d3-a456-426655440000"),
     )
-    rating: int = Field(..., description="Ratings for coffee")
+    rating: float = Field(..., description="Ratings for coffee")
     coffee_id: UUID = Field(
         ...,
         description="The id of the coffee",
